pyproject.toml
[project]
name = "simple-agent-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.0",
    "python-dotenv>=1.0.1",
    "tinker>=0.2.2",
    "torch>=2.2.0",
]

[project.scripts]
train-2048 = "simple_agent_example.train:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
